pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "screener-agent"
version = "0.1.0"
description = "Agentic system for Screener.in stock screening"
requires-python = ">=3.10"

dependencies = [
  "langchain>=0.2.0",
  "langchain-openai",
  "pydantic>=2.5",
  "playwright",
  "python-dotenv",
  "tenacity",
  "beautifulsoup4",
  "lxml",
  "fastapi",
  "uvicorn",
  "mcp"
]

[tool.setuptools]
packages = ["app"]

[tool.playwright]
browsers = ["chromium"]
